KAAV-2472 confirmation to change false if phase is deleted from timeline
